+++ packages/lit-element/src/lib/decorators/base.ts
@@ -12,28 +12,8 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-export type Constructor<T> = {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  new (...args: any[]): T;
-};
-
-// From the TC39 Decorators proposal
-export interface ClassDescriptor {
-  kind: 'class';
-  elements: ClassElement[];
-  finisher?: <T>(clazz: Constructor<T>) => undefined | Constructor<T>;
-}
-
-// From the TC39 Decorators proposal
-export interface ClassElement {
-  kind: 'field' | 'method';
-  key: PropertyKey;
-  placement: 'static' | 'prototype' | 'own';
-  initializer?: Function;
-  extras?: ClassElement[];
-  finisher?: <T>(clazz: Constructor<T>) => undefined | Constructor<T>;
-  descriptor?: PropertyDescriptor;
-}
+import {ClassElement} from 'updating-element/decorators/base.js';
+export * from 'updating-element/decorators/base.js';
 
 export const legacyPrototypeMethod = (
   descriptor: PropertyDescriptor,